pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "watchtower-cli"
version = "0.1.1"
description = "CLI tool for ingesting satellite telemetry from UDL and uploading to Nominal"
readme = "README.md"
authors = [{name = "Tarun Prakash"}]
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "nominal",
    "pandas",
    "questionary",
    "requests",
    "urllib3",
    "connect-python"
]

[tool.setuptools]
package-dir = { "" = "src" }
packages = { find = { where = ["src"], include = ["watchtower*"] } }

[tool.setuptools.package-data]
"watchtower.connect" = ["*.connect"]

[project.scripts]
watchtower = "watchtower.core.import_udl_to_nominal:main"
watchtower-connect = "watchtower.connect.cli:main"
